Feature/nn modules (#237)
Created modules structure

---------
