+++ tests/models/nn/__init__.py
@@ -11,9 +11,3 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-
-from pydantic import BaseModel
-
-
-class BaseConfig(BaseModel, extra="forbid"):
-    """Base config class for rectools."""
