+++ tests/models/test_serialization.py
@@ -1,3 +1,17 @@
+#  Copyright 2024 MTS (Mobile Telesystems)
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
 import sys
 import typing as tp
 from tempfile import NamedTemporaryFile
@@ -15,6 +29,7 @@ except ImportError:
 
 from rectools.metrics import NDCG
 from rectools.models import (
+    BERT4RecModel,
     DSSMModel,
     EASEModel,
     ImplicitALSWrapperModel,
@@ -22,11 +37,12 @@ from rectools.models import (
     LightFMWrapperModel,
     PopularInCategoryModel,
     PopularModel,
+    SASRecModel,
     load_model,
     model_from_config,
 )
 from rectools.models.base import ModelBase, ModelConfig
-from rectools.models.sasrec import SASRecModel, TransformerModelBase
+from rectools.models.nn.transformer_base import TransformerModelBase
 from rectools.models.vector import VectorModel
 
 from .utils import get_successors
@@ -49,6 +65,7 @@ CONFIGURABLE_MODEL_CLASSES = tuple(
     not in (
         DSSMModel,
         SASRecModel,
+        BERT4RecModel,
     )
 )
